Separate processing pipeline config and actions from payload functions for metrics.
