+++ devel/fetch_metrics/fetch_metrics.py
@@ -4,7 +4,8 @@ import argparse
 import json
 import os
 import time
-from typing import Any, Dict, List, TypedDict, Union, cast
+from typing import Any, Dict, List, TypedDict, cast
+from common.calling_framework import add_standard_args, parse_shared_configuration, print_per_verbose, wrap
 import spikeextractors as se
 import hither2 as hi
 import kachery_p2p as kp
@@ -16,18 +17,8 @@ expected_spiketoolkit_version = '0.7.4'
 expected_spikecomparison_version = '0.3.2'
 
 class ArgsDict(TypedDict):
-    verbose: int
-    test: int
     sortingsfile: str
     recordingset: str
-    outfile: str
-    workercount: int
-    job_cache: Union[str, None]
-    use_container: bool
-    use_slurm: bool
-    slurm_max_jobs_per_alloc: int
-    slurm_max_simultaneous_allocs: int
-    slurm_command: str
 
 RECORDING_URI_KEY = 'recordingUri'
 GROUND_TRUTH_URI_KEY = 'sortingTrueUri'
@@ -51,100 +42,33 @@ QUALITY_METRICS = [
 
 def init_args():
     parser = argparse.ArgumentParser(description="Compute ground-truth comparisons and quality metrics for SpikeForest records.")
-    parser.add_argument('--verbose', '-v', action='count', default=0)
-    parser.add_argument('--test', '-t', action='store', type=int, default=0,
-        help="If non-zero, this will set a maximum number of iterations before quitting, " +
-        "to give a usable sample without processing the entire data set.")
     parser.add_argument('--sortingsfile', '-s', action='store',
         default='sha1://31ea996f4aa43e1cb8719848753ebfed3a184503/example.json',
         help="The path or kachery URI for the JSON file which contains the sortings. The sortings file content " +
             "should be equivalent to the output of an API call to SpikeForest.")
     parser.add_argument('--recordingset', '-r', action='store', default='',
         help='If set, will limit processing to the set of recordings named in the variable (e.g. "paired_kampff").')
-    parser.add_argument('--outfile', '-o', action='store', default=None,
-        help='If set, JSON output (but not warnings/messages) will be written to this file (instead of to STDOUT).')
-    parser.add_argument('--workercount', '-w', action='store', type=int, default=4,
-        help="If set, determines the number of worker threads for a parallel job handler. Ignored if using slurm.")
-    parser.add_argument('--job_cache', action='store', type=str, default='default-job-cache',
-        help="If set, indicates the feed name for the job cache feed.")
-    parser.add_argument('--no_job_cache', action='store_true', default=False,
-        help="If set, job cache will not be used, and any value for --job_cache will be ignored.")
-    parser.add_argument('--use_container', '-C', action='store_true', default=False,
-        help='If set, hither calls will use containerization. If unset, containerization may still be used if ' +
-        'environment variable HITHER_USE_CONTAINER is set to "1" or "TRUE".')
-    parser.add_argument('--no_container', action='store_true', default=False,
-        help='Override HITHER_USE_CONTAINER environment variable to suppress container use. Ignored if --use_container is set.')
-    parser.add_argument('--use_slurm', action='store_true', default=False,
-        help='If set, this script will use a SlurmJobHandler and attempt to run jobs on the configured cluster. The exact ' +
-        'call used by the slurm job handler to acquire resources can be customized with command-line arguments.')
-    parser.add_argument('--slurm_partition', action='store', type=str, default="ccm",
-        help='If set, slurm will use the specified text as a partition name to request. Note that slurm must be explicitly ' +
-        'requested with the --use_slurm flag; if it is not, this value is ignored.')
-    parser.add_argument('--slurm_accept_shared_nodes', action='store_true', default=False,
-        help='If set, slurm calls will be made without --exclusive. Note that slurm must still be explicitly ' +
-        'requested with the --use_slurm flag; if it is not, this value is ignored.')
-    parser.add_argument('--slurm_jobs_per_allocation', action='store', type=int, default=6,
-        help='Controls the max length of job processing queues for slurm nodes. Default 6.')
-    parser.add_argument('--slurm_max_simultaneous_allocations', action='store', type=int, default=5,
-        help='The maximum number of job processing queues/slurm nodes to be requested. Default 5.')
-    parser.add_argument('--check_config', action='store_true', default=False,
-        help='Debugging tool. If set, program will simply quit with a description of the parsed configuration.')
-    parsed = parser.parse_args()
+    parsed = add_standard_args(parser)
     return parsed
 
 def init_configuration():
     args: ArgsDict = {
-        'verbose': 0,
-        'test': 0,
         'sortingsfile': '',
         'recordingset': '',
-        'outfile': '',
-        'workercount': 0,
-        'job_cache': 'default-job-cache',
-        'use_container': False,
-        'use_slurm': False,
-        'slurm_max_jobs_per_alloc': 6,
-        'slurm_max_simultaneous_allocs': 5,
-        'slurm_command': ""
     }
     parsed = init_args()
-    # We would very much like to avoid this manual copying, unfortunately the argsparse module and the typing module
-    # don't play at all nicely with each other. For a more robust script, we'd want to try a different solution,
-    # maybe the Tap/Typed Argument Parser module
-    args['verbose'] = parsed.verbose or 0
-    args['test'] = parsed.test
+    std_args = parse_shared_configuration(parsed)
     args['sortingsfile'] = parsed.sortingsfile
     args['recordingset'] = parsed.recordingset
-    args['outfile'] = parsed.outfile
-    args['workercount'] = max(parsed.workercount, 1)
-    args['job_cache'] = None if parsed.no_job_cache else parsed.job_cache
-    args['use_container'] = parsed.use_container or ((not parsed.no_container) and os.getenv('HITHER_USE_CONTAINER') in ['TRUE', '1'])
-    args['use_slurm'] = parsed.use_slurm
-    args['slurm_max_jobs_per_alloc'] = parsed.slurm_jobs_per_allocation
-    args['slurm_max_simultaneous_allocs'] = parsed.slurm_max_simultaneous_allocations
-    # example srun_command: srun --exclusive -n 1 -p <partition>
-    args['slurm_command'] = f"srun -n 1 -p {parsed.slurm_partition} {'--exclusive' if not parsed.slurm_accept_shared_nodes else ''}"
-    if parsed.outfile is not None and parsed.outfile != '' and os.path.exists(parsed.outfile) and parsed.outfile != "/dev/null":
-        raise Exception('Error: Requested to write to an existing output file. Aborting to avoid overwriting file.')
     if(parsed.check_config):
         print(f"""Received the following environment vars:
             HITHER_USE_CONTAINER: {os.getenv('HITHER_USE_CONTAINER')}
             HITHER_USE_SINGULARITY: {os.getenv('HITHER_USE_SINGULARITY')}
         """)
+        print(f"\n\tFinal Shared configuration:\n{json.dumps(std_args, indent=4)}")
         print(f"\n\tFinal configuration:\n{json.dumps(args, indent=4)}")
         exit()
-    # configure verbosity for the run
-    print_per_verbose.__dict__['verbosity_level'] = args['verbose']
-    return args
-
-def print_per_verbose(lvl: int, msg: str):
-    # verbosity_level is a static value, initialized from command-line argument at setup time in init_configuration().
-    # This does not play nicely with containerization, but global arguments variables don't either; rather than passing
-    # needless verbosity parameters around, we're just going to bail if the verbosity_level property isn't set.
-    if ('verbosity_level' not in print_per_verbose.__dict__): return
-    if (print_per_verbose.verbosity_level < lvl): return
-    tabs = max(0, lvl - 1)
-    print("\t" * tabs + msg)
+    return (args, std_args)
 
 def load_sortings(sortingsfile: str) -> List[Dict[str, Any]]:
     hydrated_sortings = kp.load_json(sortingsfile)
@@ -292,49 +216,26 @@ def output_results(comparison_list, outfile):
     else:
         print(f"Results:\n{json.dumps(comparison_list, indent=4)}")
 
+def extraction_loop(sortings, comparison_list, max_iterations = 0):
+    count = 0
+    for sorting_record in sortings:
+        print_per_verbose(2, f"Creating job-pair {count + 1} ({extract_sorting_reference_name(sorting_record)})")
+        process_sorting_record(sorting_record, comparison_list)
+        count += 1
+        if max_iterations > 0 and count >= max_iterations: break
+    print_per_verbose(1, f'{count*2} jobs have been queued. Now waiting for them to complete.')
+
 def main():
-    args = init_configuration()
-    use_container = args['use_container']
-    if args['test'] != 0: print(f"\tRunning in TEST MODE--Execution will stop after processing {args['test']} sortings!\n")
+    (args, std_args) = init_configuration()
     sortings = load_sortings(args['sortingsfile'])
     if args['recordingset'] is not None and args['recordingset'] != '':
         sortings = [s for s in sortings if s['studyName'] == args['recordingset']]
 
-    if use_container:
-        print_per_verbose(1, f"Using {'Singularity' if os.getenv('use_singularity') else 'Docker'} containers.")
-    else:
-        print_per_verbose(1, "Running without containers.")
-
-    # Define job cache and job handler
-    jc = None if args['job_cache'] == None else hi.JobCache(feed_name=args['job_cache'])
-    if args['use_slurm']:
-        jh = hi.SlurmJobHandler(
-            num_jobs_per_allocation=args['slurm_max_jobs_per_alloc'],
-            max_simultaneous_allocations=args['slurm_max_simultaneous_allocs'],
-            srun_command=args['slurm_command']
-        )
-    else:
-        jh = hi.ParallelJobHandler(num_workers=args['workercount'])
-
-    try:
-        print(f"\t\tScript execution beginning at {time.ctime()}")
-        count = 0
-        comparison_list = []
-        start_time = time.time()
-        log = hi.Log()
-        with hi.Config(job_cache=jc, job_handler=jh, use_container=use_container, log=log):
-            for sorting_record in sortings:
-                print_per_verbose(2, f"Creating job-pair {count + 1} ({extract_sorting_reference_name(sorting_record)})")
-                process_sorting_record(sorting_record, comparison_list)
-                count += 1
-                if args['test'] > 0 and count >= args['test']: break
-        print_per_verbose(1, f'{count*2} jobs have been queued. Now waiting for them to complete.')
-        hi.wait(None)
-    finally:
-        jh.cleanup()
-
-    output_results(comparison_list, args['outfile'])
-    print(f"\n\n\t\tElapsed time: {time.time() - start_time:.3f} sec")
+    comparison_list = []
+    wrap(std_args,
+        job_creation_function=extraction_loop(sortings, comparison_list, std_args["test"]),
+        post_job_fn=output_results(comparison_list, std_args['outfile'])
+    )
     print(f"\t\tScript execution complete at {time.ctime()}")
